rename analysis_widgets to idealization_frame
