+++ src/gui/__init__.py
@@ -1,7 +1,7 @@
 from .io_widgets import ExportDialog, OpenFileDialog, ExportFADialog
 from .first_activation_frame import FirstActivationFrame
 from .DISC_frame import DISCFrame
-from .analysis_widgets import IdealizationFrame
+from .idealization_frame import IdealizationFrame
 from .histogram_frame import HistogramFrame
 from .threshold_crossing_frame import ThresholdCrossingFrame
 from .processing_frames import BaselineFrame, FilterFrame
