+++ src/gui/idealization_frame.py
